+++ client/src/app/pages/dashboard/dashboard-routing.module.ts
@@ -5,6 +5,7 @@ import {DashboardHomeComponent} from './home/dashboard-home.component';
 import {UserListComponent} from './user-list/user-list.component';
 import {CommunityHomeComponent} from './community/community-home.component';
 import {NotFoundDashboardComponent} from '../not-found/dashboard/not-found-dashboard.component';
+import {UserProfileComponent} from './user-profile/user-profile.component';
 
 const routes: Routes = [
   {
@@ -22,6 +23,16 @@ const routes: Routes = [
       {
         path: 'community',
         component: CommunityHomeComponent,
+      },
+      {
+        path: 'profile',
+        component: UserProfileComponent,
+        children: [
+          {
+            path: ':id',
+            component: UserProfileComponent,
+          },
+        ],
       },
       {
         path: '',
