+++ client/src/app/pages/dashboard/user-profile/user-profile.component.scss
@@ -1,6 +1,6 @@
-@import '../../@theme/styles/themes';
-@import '~bootstrap/scss/mixins/breakpoints';
-@import '~@nebular/theme/styles/global/bootstrap/breakpoints';
+@import '../../../@theme/styles/themes';
+@import '../../../../../node_modules/bootstrap/scss/mixins/breakpoints';
+@import '../../../../../node_modules/@nebular/theme/styles/global/bootstrap/breakpoints';
 
 @include nb-install-component() {
   .nb-arrow-right {
